Create QueryForm component
