+++ src/app/components/query-form/query-form.component.css
@@ -1,0 +1,15 @@
+.grid-container {
+  margin: 20px;
+}
+
+.margin-right {
+  margin-right: 12px;
+}
+
+mgl-map {
+  width: 40vh;
+  height: 25vh;
+  min-width: 250px;
+  min-height: 155px;
+  margin-top: 10px;
+}
